Reimplemented distance culling for sparks and flashers, and added distance attenuation (i.e., they fade out instead of popping).
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@2544 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/Entities/OOFlasherEntity.m
@@ -152,7 +152,7 @@ MA 02110-1301, USA.
 - (double)findCollisionRadius
 {
 	NSSize size = [self size];
-	return fmax(size.width, size.height);
+	return fmaxf(size.width, size.height);
 }
 
 @end
